requirements.txt
django-admin-honeypot==1.1.0
django-allauth==0.38.0
django-appconf==1.0.2
django-bootstrap-form==3.4
django-bootstrap4==0.0.8
django-braces==1.13.0
django-celery-beat==2.1.0
django-channels==0.7.0
django-ckeditor==6.0.0
django-cleanup==5.2.0
django-compressor==2.2
django-cors-headers==2.4.0
django-crispy-forms==1.8.1
django-debug-toolbar==1.11
django-encrypted-model-fields==0.5.8
django-environ==0.4.5
django-extensions==3.1.3
django-filter==1.1.0
django-flatpickr==1.0.1
django-graphql-jwt==0.1.10
django-haystack==2.8.1
django-heroku==0.3.1
django-hosts==3.0
django-js-asset==1.2.2
django-markdown==0.8.4
django-model-utils==4.1.1
django-paypal==1.0.0
django-phonenumber-field==2.2.0
django-private-chat==0.2.2
django-redis==4.10.0
django-request==1.5.4
django-safedelete==0.5.1
django-softdelete==0.9.0
django-storages==1.9.1
django-tables2==2.0.4
django-taggit==1.3.0
django-timezone-field==4.2.1
django-urlcrypt==0.1.6
django-weasyprint==0.5.4
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
